s3_file to a custom resource and move modules around
Opscode::Aws is now AwsCookbook
s3_file is now a custom resource
Unnecessary libraries are now just included in the resources themselves
