+++ libraries/cloudformation.rb
@@ -1,16 +1,14 @@
 require File.join(File.dirname(__FILE__), 'ec2')
 
-module Opscode
-  module Aws
-    module CloudFormation
-      include Opscode::Aws::Ec2
+module AwsCookbook
+  module CloudFormation
+    include AwsCookbook::Ec2
 
-      def cfn
-        require 'aws-sdk'
+    def cfn
+      require 'aws-sdk'
 
-        Chef::Log.debug('Initializing the CloudFormation Client')
-        @cfn ||= create_aws_interface(::Aws::CloudFormation::Client)
-      end
+      Chef::Log.debug('Initializing the CloudFormation Client')
+      @cfn ||= create_aws_interface(::Aws::CloudFormation::Client)
     end
   end
 end
