+++ libraries/kinesis.rb
@@ -1,16 +1,14 @@
 require File.join(File.dirname(__FILE__), 'ec2')
 
-module Opscode
-  module Aws
-    module Kinesis
-      include Opscode::Aws::Ec2
+module AwsCookbook
+  module Kinesis
+    include AwsCookbook::Ec2
 
-      def kinesis
-        require 'aws-sdk'
+    def kinesis
+      require 'aws-sdk'
 
-        Chef::Log.debug('Initializing the Kinesis Client')
-        @kinesis ||= create_aws_interface(::Aws::Kinesis::Client)
-      end
+      Chef::Log.debug('Initializing the Kinesis Client')
+      @kinesis ||= create_aws_interface(::Aws::Kinesis::Client)
     end
   end
 end
